+++ Service/SiteServiceInterface.php
@@ -48,7 +48,8 @@ interface SiteServiceInterface
     /**
      * Returns an array of latest product entities
      * 
+     * @param integer $categoryId Optionally can be filtered by category id
      * @return array
      */
-    public function getLatest();
+    public function getLatest($categoryId = null);
 }